refactor to fix internal dependencies
