+++ ibft/ibft_running_instance.go
@@ -4,7 +4,7 @@ import (
 	"github.com/bloxapp/ssv/ibft/proto"
 	"github.com/bloxapp/ssv/ibft/sync/speedup"
 	"github.com/bloxapp/ssv/network"
-	"github.com/bloxapp/ssv/validator"
+	"github.com/bloxapp/ssv/utils/format"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
@@ -21,7 +21,7 @@ func (i *ibftImpl) startInstanceWithOptions(instanceOpts *InstanceOptions, value
 		return nil, errors.WithMessage(err, "could not start iBFT instance")
 	}
 
-	pk, role := validator.IdentifierUnformat(string(i.Identifier))
+	pk, role := format.IdentifierUnformat(string(i.Identifier))
 	metricsCurrentSequence.WithLabelValues(role, pk).Set(float64(i.currentInstance.State.SeqNumber.Get()))
 
 	// catch up if we can
