+++ ibft/instance.go
@@ -6,8 +6,8 @@ import (
 	"github.com/bloxapp/ssv/ibft/eventqueue"
 	"github.com/bloxapp/ssv/ibft/roundtimer"
 	"github.com/bloxapp/ssv/ibft/valcheck"
+	"github.com/bloxapp/ssv/utils/format"
 	"github.com/bloxapp/ssv/utils/threadsafe"
-	"github.com/bloxapp/ssv/validator"
 	"github.com/bloxapp/ssv/validator/storage"
 	"sync"
 	"time"
@@ -82,7 +82,7 @@ type Instance struct {
 
 // NewInstance is the constructor of Instance
 func NewInstance(opts *InstanceOptions) *Instance {
-	pk, role := validator.IdentifierUnformat(string(opts.Lambda))
+	pk, role := format.IdentifierUnformat(string(opts.Lambda))
 	metricsIBFTStage.WithLabelValues(role, pk).Set(float64(proto.RoundState_NotStarted))
 	return &Instance{
 		ValidatorShare: opts.ValidatorShare,
@@ -160,7 +160,7 @@ func (i *Instance) Start(inputValue []byte) error {
 	i.Logger.Info("Node is starting iBFT instance", zap.String("Lambda", hex.EncodeToString(i.State.Lambda.Get())))
 	i.State.InputValue.Set(inputValue)
 	i.State.Round.Set(1) // start from 1
-	pk, role := validator.IdentifierUnformat(string(i.State.Lambda.Get()))
+	pk, role := format.IdentifierUnformat(string(i.State.Lambda.Get()))
 	metricsIBFTRound.WithLabelValues(role, pk).Set(1)
 
 	if i.IsLeader() {
@@ -247,13 +247,13 @@ func (i *Instance) bumpToRound(round uint64) {
 	i.processCommitQuorumOnce = sync.Once{}
 	newRound := round
 	i.State.Round.Set(newRound)
-	pk, role := validator.IdentifierUnformat(string(i.State.Lambda.Get()))
+	pk, role := format.IdentifierUnformat(string(i.State.Lambda.Get()))
 	metricsIBFTRound.WithLabelValues(role, pk).Set(float64(newRound))
 }
 
 // ProcessStageChange set the State's round State and pushed the new State into the State channel
 func (i *Instance) ProcessStageChange(stage proto.RoundState) {
-	pk, role := validator.IdentifierUnformat(string(i.State.Lambda.Get()))
+	pk, role := format.IdentifierUnformat(string(i.State.Lambda.Get()))
 	metricsIBFTStage.WithLabelValues(role, pk).Set(float64(stage))
 
 	i.State.Stage.Set(int32(stage))
